Extract createOperationsSlice from projectStore (P5 store slice)
